ECV-593 - Resolve null materialBreakdown for scaledup bug
