+++ src/EPR.Calculator.API.BackgroundService/Exporter/CsvExporter/ScaledupProducers/CalcResultScaledupProducersExporter.cs
@@ -40,9 +40,6 @@ namespace EPR.Calculator.API.BackgroundService.Exporter.CsvExporter.ScaledupProd
             // Add data
             if (calcResultScaledupProducers.ScaledupProducers?.Any() == true)
             {
-                foreach (var producer in calcResultScaledupProducers.ScaledupProducers)
-                    producer.ScaledupProducerTonnageByMaterial = GetTonnages(producer.PomData, materials);
-
                 AppendScaledupProducers(calcResultScaledupProducers.ScaledupProducers, stringBuilder);
                 if (showTotal)
                 {
@@ -275,59 +272,6 @@ namespace EPR.Calculator.API.BackgroundService.Exporter.CsvExporter.ScaledupProd
             }
 
             return columnHeaders.ToImmutable();
-        }
-
-        public static Dictionary<string, CalcResultScaledupProducerTonnage> GetTonnages(
-            IReadOnlyCollection<ScaledupPomEntry> pomData,
-            IReadOnlyCollection<MaterialDetail> materials
-        )
-        {
-            var scaledupProducerTonnages = new Dictionary<string, CalcResultScaledupProducerTonnage>();
-
-            foreach (var material in materials)
-            {
-                var scaledupProducerTonnage = new CalcResultScaledupProducerTonnage();
-                var materialPomData = pomData.Where(e => e.MaterialId == material.Id).ToImmutableList();
-
-                var hh  = materialPomData.SingleOrDefault(e => e.PackagingType == PackagingTypes.Household);
-                var pb  = materialPomData.SingleOrDefault(e => e.PackagingType == PackagingTypes.PublicBin);
-                var cw  = materialPomData.SingleOrDefault(e => e.PackagingType == PackagingTypes.ConsumerWaste);
-
-                scaledupProducerTonnage.ReportedHouseholdPackagingWasteTonnage          = hh?.Tonnage       ?? 0;
-                scaledupProducerTonnage.ScaledupReportedHouseholdPackagingWasteTonnage  = hh?.ScaledTonnage ?? 0;
-                scaledupProducerTonnage.ReportedPublicBinTonnage                        = pb?.Tonnage       ?? 0;
-                scaledupProducerTonnage.ScaledupReportedPublicBinTonnage                = pb?.ScaledTonnage ?? 0;
-                scaledupProducerTonnage.ReportedSelfManagedConsumerWasteTonnage         = cw?.Tonnage       ?? 0;
-                scaledupProducerTonnage.ScaledupReportedSelfManagedConsumerWasteTonnage = cw?.ScaledTonnage ?? 0;
-
-                if (material.Code == MaterialCodes.Glass)
-                {
-                    var hdc = materialPomData.SingleOrDefault(e => e.PackagingType == PackagingTypes.HouseholdDrinksContainers);
-                    scaledupProducerTonnage.HouseholdDrinksContainersTonnageGlass          = hdc?.Tonnage       ?? 0;
-                    scaledupProducerTonnage.ScaledupHouseholdDrinksContainersTonnageGlass  = hdc?.ScaledTonnage ?? 0;
-
-                    scaledupProducerTonnage.TotalReportedTonnage = scaledupProducerTonnage.ReportedHouseholdPackagingWasteTonnage +
-                        scaledupProducerTonnage.ReportedPublicBinTonnage + scaledupProducerTonnage.HouseholdDrinksContainersTonnageGlass;
-                    scaledupProducerTonnage.ScaledupTotalReportedTonnage = scaledupProducerTonnage.ScaledupReportedHouseholdPackagingWasteTonnage +
-                        scaledupProducerTonnage.ScaledupReportedPublicBinTonnage + scaledupProducerTonnage.ScaledupHouseholdDrinksContainersTonnageGlass;
-                }
-                else
-                {
-                    scaledupProducerTonnage.TotalReportedTonnage = scaledupProducerTonnage.ReportedHouseholdPackagingWasteTonnage +
-                        scaledupProducerTonnage.ReportedPublicBinTonnage;
-                    scaledupProducerTonnage.ScaledupTotalReportedTonnage = scaledupProducerTonnage.ScaledupReportedHouseholdPackagingWasteTonnage +
-                        scaledupProducerTonnage.ScaledupReportedPublicBinTonnage;
-                }
-
-                scaledupProducerTonnage.NetReportedTonnage = scaledupProducerTonnage.TotalReportedTonnage -
-                    scaledupProducerTonnage.ReportedSelfManagedConsumerWasteTonnage;
-                scaledupProducerTonnage.ScaledupNetReportedTonnage = scaledupProducerTonnage.ScaledupTotalReportedTonnage -
-                    scaledupProducerTonnage.ScaledupReportedSelfManagedConsumerWasteTonnage;
-
-                scaledupProducerTonnages.Add(material.Code, scaledupProducerTonnage);
-            }
-
-            return scaledupProducerTonnages;
         }
     }
 }
