+++ src/EPR.Calculator.API.BackgroundService/Services/CalcResultReader.cs
@@ -1,10 +1,12 @@
 using EPR.Calculator.API.Data;
 using EPR.Calculator.API.Data.DataModels;
 using EPR.Calculator.API.BackgroundService.Constants;
+using EPR.Calculator.API.BackgroundService.Exporter.CsvExporter.ScaledupProducers;
 using EPR.Calculator.API.BackgroundService.Models;
 using EPR.Calculator.API.BackgroundService.Utils;
 using EPR.Calculator.API.Data.Utils;
 using Microsoft.EntityFrameworkCore;
+using EPR.Calculator.API.BackgroundService.Builder.ScaledupProducers;
 
 namespace EPR.Calculator.API.BackgroundService.Services
 {
@@ -26,7 +28,7 @@ namespace EPR.Calculator.API.BackgroundService.Services
         Task<IReadOnlyList<CalcResultCancelledProducer>> ReadCancelledProducers(int runId, CancellationToken cancellationToken);
     }
 
-    public class CalcResultReader(ApplicationDBContext dbContext) : ICalcResultReader
+    public class CalcResultReader(ApplicationDBContext dbContext, IMaterialService materialService) : ICalcResultReader
     {
         public async Task<IReadOnlyList<CalcResultH1ProjectedProducer>> ReadH1ProjectedData(int runId, CancellationToken cancellationToken)
         {
@@ -68,7 +70,7 @@ namespace EPR.Calculator.API.BackgroundService.Services
 
         public async Task<IReadOnlyList<CalcResultScaledupProducer>> ReadScaledData(int runId, CancellationToken cancellationToken)
         {
-            return await dbContext.TransformScaled
+            var scaledupProducers = await dbContext.TransformScaled
                         .Where(p => p.CalculatorRunId == runId)
                         .GroupBy(p => new { p.ProducerId, p.SubsidiaryId, p.ProducerName, p.TradingName, p.SubmissionPeriodCode, p.Level, p.IsSubTotal, p.DaysInSubmissionPeriod, p.DaysInWholePeriod, p.ScaleupFactor })
                         .Select(g =>
@@ -93,6 +95,12 @@ namespace EPR.Calculator.API.BackgroundService.Services
                         .ThenBy(p => p.SubsidiaryId)
                         .ThenBy(p => p.SubmissionPeriodCode)
                         .ToImmutableListAsync(cancellationToken);
+
+            var materials = await materialService.GetMaterials();
+            foreach (var producer in scaledupProducers)
+                producer.ScaledupProducerTonnageByMaterial = CalcResultScaledupProducersBuilder.GetTonnages(producer.PomData, materials);
+
+            return scaledupProducers;
         }
 
         public async Task<IReadOnlyList<CalcResultPartialObligation>> ReadPartialData(int runId, CancellationToken cancellationToken){
